logging_service/requirements.txt
hazelcast-python-client==4.2
flask==2.0.1
python-consul==1.1.0